+++ nextstep-jdbc/src/main/java/nextstep/jdbc/exception/IllegalConnectionException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class IllegalConnectionException extends RuntimeException {
     public IllegalConnectionException(String message) {
