+++ nextstep-jdbc/src/main/java/nextstep/jdbc/exception/IllegalExecutionException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class IllegalExecutionException extends RuntimeException {
     public IllegalExecutionException(String message) {
